rename: 일부 ui 요소 디렉토리 이동
