+++ Spender/app/src/main/java/com/example/spender/feature/home/ui/component/BubbleWithText.kt
@@ -1,9 +1,7 @@
-package com.example.spender.feature.home.ui
+package com.example.spender.feature.home.ui.component
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material3.Icon
