+++ Spender/app/src/main/java/com/example/spender/feature/home/ui/component/BudgetProgressBar.kt
@@ -1,9 +1,8 @@
-package com.example.spender.feature.home.ui
+package com.example.spender.feature.home.ui.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.offset
